fixed ba512aa from https://gitee.com/peng5233/multimedia_av_codec/pulls/1205 fix code format

+++ services/media_engine/plugins/ffmpeg_adapter/audio_decoder/ffmpeg_base_decoder.cpp
@@ -234,7 +234,7 @@ Status FfmpegBaseDecoder::ReceiveFrameSucc(std::shared_ptr<AVBuffer> &outBuffer)
                      FFMpegConverter::ConvertOHAudioChannelLayoutToString(layout).data());
         format_->SetData(Tag::AUDIO_CHANNEL_LAYOUT, layout);
     }
-    outBuffer->pts_ = static_cast<uint64_t>(cachedFrame_->pts);
+    outBuffer->pts_ = cachedFrame_->pts;
     ioInfoMem->SetSize(outputSize);
     return Status::OK;
 }
